app.py

# -------- Database Connection --------
import streamlit as st
import mysql.connector
import hashlib
import io
from PIL import Image
import pandas as pd
from fpdf import FPDF
import base64

def connect_db():
    return mysql.connector.connect(
        host=st.secrets["mysql"]["host"],
        user=st.secrets["mysql"]["user"],
        password=st.secrets["mysql"]["password"],
        database=st.secrets["mysql"]["database"]
    )

# (Rest of your code here - skipping for brevity)
st.title("✅ Deployment Successful! Modify app.py as per your project.")
