requirements.txt
streamlit
mysql-connector-python
Pillow
pandas
fpdf
